api/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
dnspython>=2.4.2
whois>=0.9.27
concurrent-log-handler>=0.9.24
colorama>=0.4.6
pyjwt>=2.8.0
flask>=3.0.0
flask-cors>=4.0.0
cryptography>=41.0.7
pyyaml>=6.0.1
jinja2>=3.1.2
tqdm>=4.66.1
shodan>=1.31.0
googlesearch-python>=1.2.3
python-dotenv>=1.0.0
censys>=2.2.8
builtwith>=1.3.4
fake-useragent>=1.4.0
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
cryptography>=41.0.7
pyopenssl>=23.0.0
redis>=5.0.1
cachetools>=5.3.2
requests>=2.31.0
python-nmap>=0.6.1
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.11.0
pylint>=3.0.2
python-json-logger>=2.0.7
prometheus-flask-exporter>=0.22.4
pyjson>=1.3.0
theHarvester
hunter>=3.3.4
sublist3r>=1.0.0
requests-html>=0.10.0
passivetotal>=2.5.0
virustotal-python>=1.0.0
spyse-python>=1.0.0
colorama>=0.4.6
python-nmap>=0.6.1
scapy>=2.4.5
pyopenssl>=23.0.0
paramiko>=2.12.0
pymongo>=4.3.3
redis>=5.0.1
pyodbc>=4.0.32
flask-limiter>=3.5.0
gunicorn>=21.2.0
pytest>=7.4.3
pytest-cov>=4.1.0
mypy>=1.7.1
black>=23.11.0
pylint>=3.0.2
safety>=2.3.5
bandit>=1.7.5
python-json-logger>=2.0.7
cachetools>=5.3.2
prometheus-flask-exporter>=0.22.4
openai>=1.0.0
google-generativeai>=0.3.0
groq>=0.3.0
anthropic>=0.8.0
flask-session>=0.5.0
requests-oauthlib>=1.3.1
websockets==11.0.3
ptyprocess==0.7.0
psutil==5.9.5